Backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pytest>=8.0.0
httpx>=0.26.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
